now builds on windows, doesn't yet pass regression test. probably build options or malloc issues

+++ testsuite/allocatePopExternal.cpp
@@ -62,7 +62,7 @@ public:
         auto &mem( output[ "y" ].allocate< obj_t >() );
         for( auto i( 0 ); i < mem.length; i++ )
         {
-            mem.pad[ i ] = counter;
+            mem.pad[ i ] = static_cast< int >( counter );
         }
         output[ "y" ].send();
         counter++;
